requirements.txt
gunicorn==19.9.0
python-telegram-bot==12.2.0
python-dotenv==0.10.3
sqlalchemy==1.3.10
psycopg2==2.8.4